modified task tracker tutorial files to be more object-oriented
