+++ examples/react-crash-course/react-task-tracker/src/components/Header.js
@@ -1,23 +1,15 @@
 import PropTypes from 'prop-types'
-import Button from './Button'
 
 
 
 // Allow people to give props to the components so they can be customized 
 // const Header = ({ title }) => {      // This will also work, but I think it's harder to read
-const Header = ({ headerTitle, headerBtnText } ) => {
-  // This is just here temporarily. We will move this later.
-  const onClickFunction = () => {
-    console.log('Click')
-  }
+const Header = ({ headerTitle } ) => {
+  
 
   return (
     <header className='header'>
       <h1>{headerTitle}</h1>
-      <Button 
-        text={headerBtnText}
-        onClick={onClickFunction}
-      />
     </header>
   )
 }
@@ -25,7 +17,6 @@ const Header = ({ headerTitle, headerBtnText } ) => {
 // We can set default props below the component
 Header.defaultProps = {
   headerTitle: 'defaultProp Title',
-  headerBtnText: 'defaultProp Btn Text',
 }
 
 Header.propTypes = {
